+++ components/constants.js
@@ -1,0 +1,4 @@
+export const formTypes = {
+  contact: 'contact',
+  question: 'question',
+};
